+++ src/auth0__userinfo__/index.d.ts
@@ -1,10 +1,8 @@
 import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be, Ctx } from '@ctx-core/object'
-import type { auth0__userinfo_T } from '../auth0__userinfo__fetch_get'
+import type { Be } from '@ctx-core/object'
+import type { auth0__userinfo_T } from '../_types'
 export declare const auth0__userinfo__:Be<ReadableAtom_<auth0__userinfo_T>>
 export {
 	auth0__userinfo__ as auth0_userinfo__,
 	auth0__userinfo__ as auth0_userinfo$_,
 }
-export function auth0__userinfo_(ctx:Ctx):auth0__userinfo_T
-export function auth0__userinfo__set(ctx:Ctx, auth0__userinfo:auth0__userinfo_T):void