+++ src/auth0__userinfo__/index.js
@@ -1,58 +1,22 @@
-import { waitfor_fibonacci_backoff } from '@ctx-core/fetch-undici'
+import { nullish__check_ } from '@ctx-core/function'
 import { be_ } from '@ctx-core/object'
-import { setter_computed_ } from '@ctx-core/nanostores'
-import { AUTH0_DOMAIN__ } from '../AUTH0_DOMAIN__/index.js'
-import { auth0__token__ } from '../auth0__token__/index.js'
-import { auth0__userinfo_token__ } from '../auth0__userinfo_token__/index.js'
-import { auth0__token__clear } from '../auth0__token__clear/index.js'
-import { auth0__userinfo__fetch_get } from '../auth0__userinfo__fetch_get/index.js'
+import { computed_ } from '@ctx-core/nanostores'
+import { auth0__userinfo__response_pair__ } from '../auth0__userinfo__response_pair__/index.js'
 /** @typedef {typeof import('@ctx-core/object').Ctx}Ctx */
 /** @typedef {typeof import('../auth0__userinfo__fetch_get').auth0__userinfo_T}auth0__userinfo_T */
 /** @type {typeof import('./index.d.ts').auth0__userinfo__} */
 export const auth0__userinfo__ = be_('auth0__userinfo__', ctx=>
-	setter_computed_([
-		AUTH0_DOMAIN__(ctx), auth0__token__(ctx), auth0__userinfo_token__(ctx),
-	], ([AUTH0_DOMAIN, auth0__token, auth0_userinfo_auth0_token,], set)=>{
-		if (auth0__token === auth0_userinfo_auth0_token) {
-			return
-		}
-		if (!auth0__token) {
-			set(auth0__token === undefined ? undefined : null)
-			return
-		}
-		set(auth0__token)
-		/** @type {import('./auth0__userinfo__fetch_get.d.ts').auth0__userinfo__fetch_get__payload_T} */
-		let auth0_userinfo
-		waitfor_fibonacci_backoff(
-			async ()=>{
-				const [_auth0_userinfo, response] = await auth0__userinfo__fetch_get({ auth0__token, AUTH0_DOMAIN })
-				auth0_userinfo = _auth0_userinfo
-				return response
-			}
-		).then(response=>{
-			if (!response.ok) {
-				auth0__token__clear(ctx)
-				set(null)
-				return
-			}
-			set(auth0_userinfo)
-		})
-	}))
+	computed_(
+		auth0__userinfo__response_pair__(ctx),
+		auth0__userinfo__response_pair=>
+			nullish__check_([auth0__userinfo__response_pair], ()=>{
+				const [payload] = auth0__userinfo__response_pair
+				return (
+					typeof payload === 'string' || 'error' in payload
+					? null
+					: payload)
+			})))
 export {
 	auth0__userinfo__ as auth0_userinfo__,
 	auth0__userinfo__ as auth0_userinfo$_,
-}
-/**
- * @param ctx{Ctx}
- * @returns {auth0__userinfo_T}
- */
-export function auth0__userinfo_(ctx) {
-  return auth0__userinfo__(ctx).$
-}
-/**
- * @param ctx{Ctx}
- * @param auth0__userinfo{auth0__userinfo_T}
- */
-export function auth0__userinfo__set(ctx, auth0__userinfo) {
-	auth0__userinfo__(ctx).$ = auth0__userinfo
 }