+++ src/auth0__userinfo__response_pair__/index.d.ts
@@ -1,0 +1,11 @@
+import type { Be } from '@ctx-core/object'
+import type { Ctx } from '@ctx-core/object'
+import type { ReadableAtom_ } from '@ctx-core/nanostores'
+import type { auth0__userinfo__response_pair_T } from '../_types'
+export declare const auth0__userinfo__response_pair__:Be<
+	ReadableAtom_<auth0__userinfo__response_pair_T>
+>
+export function auth0__userinfo__response_pair_(ctx:Ctx):auth0__userinfo__response_pair_T
+export function auth0__userinfo__response_pair__set(
+	ctx:Ctx, auth0__userinfo__response_pair:auth0__userinfo__response_pair_T
+):void